+++ app/parsers/__init__.py
@@ -1,0 +1,2 @@
+from app.parsers.dspipelines import *
+from app.parsers._factory import *