+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Parser metadata in JSON ML profile.
+- "dspipelines" parser
+
 ## [0.3] - 2025-07-16
 
 ### Added